// packages/protocol/typescript/generated/yiru/runtime/v1/git_common_pb.d.ts
export enum GitRefCategory {
  /**
   * @generated from enum value: GIT_REF_CATEGORY_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: GIT_REF_CATEGORY_BRANCH = 1;
   */
  BRANCH = 1,

  /**
   * @generated from enum value: GIT_REF_CATEGORY_REMOTE_BRANCH = 2;
   */
  REMOTE_BRANCH = 2,

  /**
   * @generated from enum value: GIT_REF_CATEGORY_TAG = 3;
   */
  TAG = 3,

  /**
   * @generated from enum value: GIT_REF_CATEGORY_COMMIT = 4;
   */
  COMMIT = 4,

  /**
   * @generated from enum value: GIT_REF_CATEGORY_HEAD = 5;
   */
  HEAD = 5,
}